requirements.txt

fastapi==0.89.1
uvicorn==0.20.0
configparser
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]